V1.5 update
Major:
Added ability to upload images to the Q&A
Made the Q&A section stateful-messages are passed to it as a propper  conversation thread [{},{},{},{}] instaed of a simple system and user message [{},{}].
Cached conversation history
Minor changes:
disabled web search for Q&A as results were irelevant.
Changed ask your own question box to be a textarea for wrapability.
Tweaked LLM system instructions and response templates.
Bumped version to 1.5